fix: Use `modules` instead of `pages`
