+++ apps/flame-defi/app/earn/(routes)/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Page as VaultListPage } from "earn/pages/vault-list";
+import { Page as VaultListPage } from "earn/modules/vault-list";
 
 export default function Page() {
   return <VaultListPage />;
