+++ apps/flame-defi/app/earn/modules/vault-list/components/table/table-search.tsx
@@ -1,7 +1,7 @@
 import { SearchIcon } from "@repo/ui/icons";
 import { cn } from "@repo/ui/lib";
 import { Input, Skeleton } from "@repo/ui/shadcn-primitives";
-import { usePageContext } from "earn/pages/vault-list/hooks/usePageContext";
+import { usePageContext } from "earn/modules/vault-list/hooks/usePageContext";
 import { useRef, useState } from "react";
 
 export const TableSearch = () => {
