+++ apps/flame-defi/app/earn/modules/vault-list/components/table/table.tsx
@@ -3,7 +3,7 @@ import { cn } from "@repo/ui/lib";
 import { Skeleton } from "@repo/ui/shadcn-primitives";
 import { flexRender } from "@tanstack/react-table";
 import { OrderDirection } from "earn/gql/graphql";
-import { usePageContext } from "earn/pages/vault-list/hooks/usePageContext";
+import { usePageContext } from "earn/modules/vault-list/hooks/usePageContext";
 
 export const Table = () => {
   const {
